requirements.txt
streamlit>=1.25.0
langchain>=0.3.0
openai>=1.58.1
python-dotenv